Fluids - remove blasius_exact.h
