+++ examples/fluids/qfunctions/blasius.h
@@ -16,8 +16,62 @@
 #include "newtonian_state.h"
 #include "newtonian_types.h"
 #include "utils.h"
-#include "blasius_exact.h"
 
+typedef struct BlasiusContext_ *BlasiusContext;
+struct BlasiusContext_ {
+  bool       implicit; // !< Using implicit timesteping or not
+  bool       weakT;    // !< flag to set Temperature weakly at inflow
+  CeedScalar delta0;   // !< Boundary layer height at inflow
+  CeedScalar Uinf;     // !< Velocity at boundary layer edge
+  CeedScalar Tinf;     // !< Temperature at boundary layer edge
+  CeedScalar T_wall;   // !< Temperature at the wall
+  CeedScalar P0;       // !< Pressure at outflow
+  CeedScalar theta0;   // !< Temperature at inflow
+  CeedScalar x_inflow; // !< Location of inflow in x
+  CeedScalar n_cheb;   // !< Number of Chebyshev terms
+  CeedScalar *X;       // !< Chebyshev polynomial coordinate vector
+  CeedScalar eta_max;  // !< Maximum eta in the domain
+  CeedScalar *Tf_cheb; // !< Chebyshev coefficient for f
+  CeedScalar *Th_cheb; // !< Chebyshev coefficient for h
+  struct NewtonianIdealGasContext_ newtonian_ctx;
+};
+
+// *****************************************************************************
+// This helper function evaluates Chebyshev polynomials with a set of
+//  coefficients with all their derivatives represented as a recurrence table.
+// *****************************************************************************
+CEED_QFUNCTION_HELPER void ChebyshevEval(int N, const double *Tf, double x,
+    double eta_max, double *f) {
+  double dX_deta   = 2 / eta_max;
+  double table[4][3] = {
+    // Chebyshev polynomials T_0, T_1, T_2 of the first kind in (-1,1)
+    {1, x, 2*x *x - 1}, {0, 1, 4*x}, {0, 0, 4}, {0, 0, 0}
+  };
+  for (int i=0; i<4; i++) {
+    // i-th derivative of f
+    f[i] = table[i][0] * Tf[0] + table[i][1] * Tf[1] + table[i][2] * Tf[2];
+  }
+  for (int i=3; i<N; i++) {
+    // T_n(x) = 2xT_{n-1}(x) - T_{n-2}(x)
+    table[0][i%3] = 2 * x * table[0][(i-1) % 3] - table[0][(i-2)%3];
+    // Differentiate Chebyshev polynomials with the recurrence relation
+    for (int j=1; j<4; j++) {
+      // T'_{n}(x)/n = 2T_{n-1}(x) + T'_{n-2}(x)/n-2
+      table[j][i%3] = i * (2 * table[j-1][(i-1) % 3] + table[j][(i-2)%3] / (i-2));
+    }
+    for (int j=0; j<4; j++) {
+      f[j] += table[j][i%3] * Tf[i];
+    }
+  }
+  for (int i=1; i<4; i++) {
+    // Transform derivatives from Chebyshev [-1, 1] to [0, eta_max].
+    for (int j=0; j<i; j++) f[i] *= dX_deta;
+  }
+}
+
+// *****************************************************************************
+// This helper function computes the Blasius boundary layer solution.
+// *****************************************************************************
 State CEED_QFUNCTION_HELPER(BlasiusSolution)(const BlasiusContext blasius,
     const CeedScalar x[3], const CeedScalar x0, const CeedScalar x_inflow,
     const CeedScalar rho, CeedScalar *t12) {
